+++ sand/src/tracer.rs
@@ -68,8 +68,8 @@ impl<'t, F: Future<Output = ()>> Tracer<'t, F> {
                 self.as_mut().message_event(message);
             }
             match ptrace::wait(&mut siginfo) {
-                err if err == -abi::ECHILD => break,
-                err if err == -abi::EINTR => (),
+                err if err == -abi::ECHILD as isize => break,
+                err if err == -abi::EINTR as isize => (),
                 err if err == 0 => self.as_mut().siginfo_event(&siginfo),
                 err => panic!("unexpected waitid response ({})", err),
             }
